Livestream-only query cleanup.
This completes the change in c24c016e, covering ClaimListDiscover as well.
